Refactored zero rated controller package
